+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 [Unreleased]: https://github.com/chaosiq/chaosiq-cloud/compare/0.9.5...HEAD
 
+### Changed
+
+-  Add prometheus greater/lower tolerances
+
 ## [0.9.5][] - 2020-03-05
 
 [0.9.5]: https://github.com/chaosiq/chaosiq-cloud/compare/0.9.4...0.9.5